mcp-server/requirements.txt
mcp==0.1.0
sqlalchemy==2.0.25
asyncpg==0.29.0
pydantic==2.5.3
python-dotenv==1.0.0
